requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
fake-useragent>=1.4.0
PyYAML>=6.0
selenium>=4.15.0
webdriver-manager>=4.0.0
tqdm>=4.66.0